DailyF303Generator cleaned
